+++ contrib/prune/generic_list.h
@@ -10,15 +10,17 @@
 
 #pragma once
 
+#include <stddef.h>
+
     typedef void *gl_data;
 
     typedef struct {
-	uint64_t used;	/* number of elements in the list */
-	uint64_t size;	/* number of elements that the list can hold */
+	size_t used;	/* number of elements in the list */
+	size_t size;	/* number of elements that the list can hold */
 	gl_data *data;		/* pointer to first element */
     } generic_list_t;
 
-    extern generic_list_t new_generic_list(uint64_t size);
+    extern generic_list_t new_generic_list(size_t size);
     extern void add_to_generic_list(generic_list_t * list,
 					       gl_data element);
     extern void free_generic_list(generic_list_t * list);
